+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Added
+
+- `AvatarCombo`, `AvatarIcon` & `AvatarUser` now supports `role="button"`
+  - Added support for common DOM properties and event handlers (e.g.: `onClick`)
+  - Improved a11y for all `Avatar*` components
+  - Added Storybook with knobs
+
 ### Changed
 
 - `theme` "pressed" colors are more discernable from other stateful colors